hack-latex, Makefile: Hacking to make internal cross-references work
References between chapters use Markdown filenames and (except at
chapter level) fragment identifiers.  Pandoc mostly puts in helpful
labels which match the fragment names, except that they're based on the
section titles, and not unique across the entire document.  The labels
for chapters are less helpful because they're only actually referred to
using filenames, not the chapter titles.

Introduce a Perl script which hacks all of the generated LaTeX labels
and internal references.  The labels are all qualified with the
originating filename to uniquify them, and chapter labels are now /just/
the filename, so references from other files work easily.  References to
fragments are fixed up.

The Perl script will later also bodge the overly-wide tables.  But not
yet.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>