src/polyglot.tex: Fix the running headers and footers
The default LaTeX page styles have an awful defect: the `plain' style,
as used on the first page of a chapter, is entirely different from the
`headings' style used on ordinary pages -- the page number is at the
centre of the footer, rather than the outer side of the header -- which
makes flipping through a book to find a page just that bit more
annoying.

Overwrite the page styles to do a better job.  Page numbers are always
in the footer, at the outer edge; the header shows chapter and section,
except on chapter starts, cleared pages, and in the front matter.  (This
last was an awful job to achieve because `memoir' sets the
`\if@mainmatter' switch /before/ clearing the page, so the last page of
the Table of Contents ends up with a spurious header.  Fix by clearing
by hand before switching to the `\mainmatter' division.)

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>