src/polyglot.tex: Overhaul the page layout
This is taken from my `strayman' document class.  The text block is in
the same place on both recto and verso pages, which makes it rather
easier on the eyes when paging through a PDF on-screen (as I expect most
readers to do).  This design consistently biases the text to the right,
leaving a large left margin, which can be occupied by headers, footers,
chapter titles and section numbers (not implemented yet), and marginal
notes (if any).  The border around the main page area is uniform on all
sides.

The previous design left a large margin on the outside edge of the
paper, and a surprisingly large gap at the bottom of the page.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>