generate-inputs: Abstract out the `fancy-pre' markup hacking
This is now done by a new function `translate_fancypre_markup', which is
given a hash(ref) describing the resulting output markup
translations.  (There's only one so this is trivial.)

The intent here is to avoid clobbering the focus line with
output-specific markup, so that we can produce half-decent Pandoc/LaTeX
output along the way.  Of course, since we split the line into left and
right halves, we now need to apply the same markup translations to both
halves.

There's no intentional change in overall output at this time.

Signed-off-by: Mark Wooding <mdw@distorted.org.uk>